Enabling linter for the OSPO book and fixing linting errors
This enables and configures a GitHub action to run a markdown linter
against the chapters of the OSPO book. This also fixes a range of
markdown syntax errors to make the linter happy.

The scope of the linter may be extended to other markdown files in
the future, once it has proven to be useful for the authors.

Signed-off-by: Georg Kunz <georg.kunz@ericsson.com>