Add some useful info and links about security fixes #566
I think the OSPO book should at least briefly note that
security is important and point to some useful resources for
more information. Here's my attempt at that.

We don't have the URL quite yet for the CRA course, but we
know its name and identifier, so including that should be adequate
for now until we have the URL.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>